# Fix overflow when rounding