This PR aims to fix #37171. It simply uses a smaller step (square root of it) twice if it overflows. Otherwise, the computation remains exactly the same as before (and thus should not break any code).

Cc to the original author: @simonbyrne

The performance remains almost the same. Tested on i5-1035G1 by
```julia
for dig=1:7
    @btime round($(rand()), sigdigits = $dig)
end
```
Master:
```
  13.995 ns (0 allocations: 0 bytes)
  14.338 ns (0 allocations: 0 bytes)
  14.582 ns (0 allocations: 0 bytes)
  23.895 ns (0 allocations: 0 bytes)
  125.770 ns (0 allocations: 0 bytes)
  125.735 ns (0 allocations: 0 bytes)
  125.771 ns (0 allocations: 0 bytes)
```
This PR:
```
  14.233 ns (0 allocations: 0 bytes)
  14.536 ns (0 allocations: 0 bytes)
  15.044 ns (0 allocations: 0 bytes)
  24.183 ns (0 allocations: 0 bytes)
  126.037 ns (0 allocations: 0 bytes)
  126.018 ns (0 allocations: 0 bytes)
  126.134 ns (0 allocations: 0 bytes)
```

(The performance drop for 5+ digits is caused by the exponent while computing the step, see #42031)